Extract dataset models from `models/project.py` into `models/dataset.py` (close #870)
