+++ fractal_server/app/models/__init__.py
@@ -2,6 +2,7 @@
 `models` module
 """
 from ..schemas import *  # noqa F401
+from .dataset import *  # noqa: F403, F401
 from .job import *  # noqa: F403, F401
 from .project import *  # noqa: F403, F401
 from .security import *  # noqa: F403, F401
